added changes to user query process and separated sentiment class
