+++ baseline/generator/generator.py
@@ -5,11 +5,10 @@ import ollama
 import os
 from transformers import AutoTokenizer, AutoModelForSequenceClassification
 from transformers import pipeline
-from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer
 from collections import defaultdict, Counter
 import torch
 import numpy as np
-import re
+import sys
 # import nltk
 
 # nltk.download("vader_lexicon")
@@ -23,6 +22,9 @@ logging.basicConfig(filename=log_path, level=logging.INFO, format="%(message)s")
 # Path to the chunked reviews file  os.path.join('..', 'data', 'reviews.json')
 project_root = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", ".."))
 chunked_file = os.path.join(project_root,"baseline", "data", "reviews.json")
+
+sys.path.append(os.path.abspath(os.path.join("..", "sentiment_analysis")))
+from sentiment_analyzer import SentimentAnalyzer
 
 """
 This generator module is designed to:
@@ -38,16 +40,9 @@ class Generator:
         self.chunk_file = chunk_file
         self.max_tokens = 300
         self.model_name = "gemma2:2b"
-
+        self.sentiment_analyzer = SentimentAnalyzer()
         with open(self.chunk_file, "r", encoding="utf-8") as f:
             self.chunked_data = json.load(f)
-
-        # Initialize the Twitter RoBERTa sentiment model
-        self.sentiment_model_name = "cardiffnlp/twitter-roberta-base-sentiment"
-        self.tokenizer = AutoTokenizer.from_pretrained(self.sentiment_model_name)
-        self.model = AutoModelForSequenceClassification.from_pretrained(self.sentiment_model_name)
-        self.device = torch.device("mps" if torch.backends.mps.is_available() else "cpu")
-        self.model.to(self.device)
 
         self.aspect_keywords = {
             "battery": ["battery", "charge", "charging", "mah", "power", "drain"],
@@ -58,14 +53,6 @@ class Generator:
             "software": ["ui", "os", "update", "bloatware", "interface", "android", "software"],
             "heating": ["heat", "heating", "warm", "temperature", "overheat"]
         }
-
-    def neg_count(self, reviews_by_sentiment):
-        neg_count = len(reviews_by_sentiment.get("negative", [])) if reviews_by_sentiment else 0
-        print('neg_count',neg_count)
-        total_count = sum(len(v) for v in reviews_by_sentiment.values()) if reviews_by_sentiment else len(review_list)
-        neg_pct = (neg_count / total_count) * 100 if total_count else 0
-        return neg_pct
-
     """
     This method creates a prompt for the Gemma model to summarize customer feedback.
     It formats the user query and the list of reviews into a structured prompt.
@@ -78,7 +65,7 @@ class Generator:
 
     def create_gemma_prompt(self, user_query, review_list, reviews_by_sentiment=None, aspect=None, sentiment_block=None):
         all_reviews_text = "\n".join(f"- {sentence}" for sentence in review_list)
-        neg_pct = self.neg_count(reviews_by_sentiment)
+        neg_pct = self.sentiment_analyzer.neg_count(reviews_by_sentiment, review_list)
         neg_instruction = "" # should we add something in the default ? #TODO
         if neg_pct < 20:
             neg_instruction = (
@@ -118,7 +105,7 @@ class Generator:
         if not review_list:
             return "No relevant reviews found."
         
-        sentiment_block, reviews_by_sentiment, aspect_scores = self.analyze_sentiment(review_list, aspect)
+        sentiment_block, reviews_by_sentiment, aspect_scores = self.sentiment_analyzer.analyze_sentiment(review_list, aspect)
 
         prompt = self.create_gemma_prompt(user_query, review_list, reviews_by_sentiment, aspect, sentiment_block)
 
@@ -148,127 +135,6 @@ class Generator:
             f.write("\n")
         return final_summary, aspect_scores
     
-    """
-    This method performs sentiment analysis on a list of reviews using a transformer-based 
-    sentiment classification model. It optionally filters sentiment analysis by a specific aspect 
-    (e.g., battery, camera). If no aspect is provided, it maps sentiments to detected aspects 
-    using predefined keywords.      
-    """
-    def analyze_sentiment(self, review_list, aspect=None):
-        sentiment_analyzer = pipeline("sentiment-analysis", model=self.model, tokenizer=self.tokenizer, device=0 if torch.cuda.is_available() else -1)
-        aspects = list(self.aspect_keywords.keys())
-        aspect_sentiments = defaultdict(list) 
-        general_counts = Counter()
-        reviews_by_sentiment = {"positive": [], "negative": [], "neutral": []}
-
-        for review in review_list:
-            review_text = review["text"] if isinstance(review, dict) and "text" in review else review
-            review_text_lower = review_text.lower()
-            # tokenize review into words (simple split on non-word chars)
-            tokens = set(re.findall(r'\b\w+\b', review_text_lower))
-
-            result = sentiment_analyzer(review_text)[0]
-            label_map = {"LABEL_0": "negative", "LABEL_1": "neutral", "LABEL_2": "positive"}
-            label = label_map.get(result["label"], "neutral")
-
-            reviews_by_sentiment[label].append(review_text)
-
-            if aspect:
-                if any(kw in tokens for kw in self.aspect_keywords.get(aspect, [])):
-                    general_counts[label] += 1
-            else:
-                matched = False
-                for asp, keywords in self.aspect_keywords.items():
-                    if any(kw in tokens for kw in keywords):
-                        aspect_sentiments[label].append(asp)
-                        matched = True
-                        break
-                if not matched:
-                    aspect_sentiments[label].append("general")
-
-        def calculate_percentage_breakdown(pos, neg, neu):
-            total = pos + neg + neu
-            if total == 0:
-                return {"positive": 0, "negative": 0, "neutral": 0}
-
-            raw = {
-                "positive": (pos / total) * 100,
-                "negative": (neg / total) * 100,
-                "neutral":  (neu / total) * 100
-            }
-
-            rounded = {k: round(v) for k, v in raw.items()}
-            diff = 100 - sum(rounded.values())
-
-            # Fix rounding difference
-            if diff != 0:
-                # Sort by decimal difference to adjust the closest
-                adjust_order = sorted(raw, key=lambda k: raw[k] - rounded[k], reverse=(diff > 0))
-                for i in range(abs(diff)):
-                    rounded[adjust_order[i % 3]] += 1 if diff > 0 else -1
-
-            return rounded
-        
-        total_reviews = len(review_list)
-
-        if aspect:
-            pos = general_counts["positive"]
-            neg = general_counts["negative"]
-            neu = general_counts["neutral"]
-
-            percentages = calculate_percentage_breakdown(pos, neg, neu)
-            overall = max(percentages, key=percentages.get).capitalize()
-
-            pos = [aspect] if general_counts["positive"] > 0 else []
-            neg = [aspect] if general_counts["negative"] > 0 else []
-            neu = [aspect] if general_counts["neutral"] > 0 else []
-            aspect_scores = self.convert_to_aspect_scores(percentages, pos, neg, neu)
-            sentiment_block = (
-                f"Overall Sentiment : {overall}\n"
-                f"- {percentages['positive']}% of customers have positive reviews.\n"
-                f"- {percentages['negative']}% of customers have negative reviews.\n"
-                f"- {percentages['neutral']}% of customers have neutral reviews.\n\n"
-            )
-        else:
-            pos_aspects = [a for a in aspect_sentiments["positive"] if a != "general"]
-            neg_aspects = [a for a in aspect_sentiments["negative"] if a != "general"]
-            neu_aspects = [a for a in aspect_sentiments["neutral"] if a != "general"]
-
-            pos = len(aspect_sentiments["positive"])
-            neg = len(aspect_sentiments["negative"])
-            neu = len(aspect_sentiments["neutral"])
-
-            percentages = calculate_percentage_breakdown(pos, neg, neu)
-            overall = max(percentages, key=percentages.get).capitalize()
-            pos = [aspect] if general_counts["positive"] > 0 else []
-            neg = [aspect] if general_counts["negative"] > 0 else []
-            neu = [aspect] if general_counts["neutral"] > 0 else []
-            aspect_scores = self.convert_to_aspect_scores(percentages, pos_aspects, neg_aspects, neu_aspects)
-            sentiment_block = (
-                f"OVERALL SENTIMENT : {overall}\n"
-                f"- {percentages['positive']}% of customers have positive reviews about {', '.join(sorted(set(pos_aspects))) or 'various aspects'}.\n"
-                f"- {percentages['negative']}% have negative reviews about {', '.join(sorted(set(neg_aspects))) or 'various aspects'}.\n"
-                f"- {percentages['neutral']}% have neutral reviews about {', '.join(sorted(set(neu_aspects))) or 'various aspects'}.\n\n"
-            )
-        print("Sentiment Block:", sentiment_block)
-        print("Reviews by Sentiment:", reviews_by_sentiment)
-        print("Aspect Scores:", aspect_scores)
-        
-        return sentiment_block, reviews_by_sentiment, aspect_scores
-
-    def convert_to_aspect_scores(self, percentages, pos_aspects, neg_aspects, neu_aspects):
-        aspect_scores = defaultdict(lambda: {"positive": 0, "neutral": 0, "negative": 0})
-        for aspect in pos_aspects:
-            aspect_scores[aspect]["positive"] = percentages["positive"]
-        for aspect in neg_aspects:
-            aspect_scores[aspect]["negative"] = percentages["negative"]
-        for aspect in neu_aspects:
-            aspect_scores[aspect]["neutral"] = percentages["neutral"]
-        
-        aspect_scores = json.dumps(aspect_scores, indent=2)
-        print("Aspect Scores:", aspect_scores)
-        return aspect_scores
-
 
 if __name__ == "__main__":
     generator = Generator()
